Refactor app initialization for i18n
Make sure the locale data is setup before importing anything, since block names
get translated on import
